Bump version to 0.9.4
We jumped 0.9.3 because it will be a bugfixes-update over 0.9.2.

Alot changes have happened since 0.9.2 (which would be scheduled for 0.9.3) but due to some people already using 0.9.2 I decided to backport only the bugfixes, call them 0.9.3 and bump the really new code to 0.9.4! (for safety and stability)

Hope it makes sense! :+1: