Add a PluginAPI Framework
This framework will be packaged with ManageConky and everyone will be able to use it when building for ManageConky.

It will provide access to basic API features.
(Probably later, there will be support for subclasssing and reimplementing some basic functionality;

for example you will be able to override the Widget/Themes search algorithm.)